Add ImageTool to editor and enhance blog workflow
Refactor redirection logic in blog controller to update redirect URLs for creation and editing processes, improving clarity. Adjust form validation formatting for better readability. In `BlogCreate.js`, optimize reset behavior based on edit mode, enhancing user experience during updates. Introduce smooth scrolling to top on success.

Integrate `@editorjs/image` into the editor with configuration for file and URL uploads, replacing `SimpleImage`. Update `package.json` to include the new dependency.

+++ assets/admin/pages/blogs/BlogCreate.js
@@ -179,14 +179,28 @@ export default class extends GiaComponent {
 
         Btn.loading(FormBtn(target))
 
+        // @ts-expect-error
+        const editMode = target.action.includes('blogs/update')
+
         // @ts-ignore
         ApiRequest('post', target.action, datas, true)
             .then(({ data: { message, redirect_url } }) => {
+                window.setTimeout(() => {
+                    window.scrollTo({ top: 0, left: 0, behavior: "smooth" })
+                }, 1000)
+
                 Notifier.success(message)
-                    .then(() => redirect_url && $swup.loadPage({ url: redirect_url }, true))
-                this.editor.clear()
-                this.clearInput(target)
-                this.pond && this.pond.removeFile(0)
+                    .then(() => {
+                        if(redirect_url) {
+                            $swup.loadPage({ url: redirect_url }, true)
+                        }
+                    })
+
+                if(!editMode) {
+                    this.editor.clear()
+                    this.clearInput(target)
+                    this.pond && this.pond.removeFile(0)
+                }
             })
             .finally(() => Btn.hide())
     }
